test: migrate component tests to the story gallery
Playwright no longer recommends @playwright/experimental-ct-*; the
supported approach is a story gallery served by the project's own dev
server, driven by the built-in mount() fixture.

- add a Vite dev server and a gallery at playwright/gallery that resolves
  stories via import.meta.glob and renders them into #root, reusing the
  root so update() reconciles instead of remounting
- move the components under test into playwright/stories/*.story.tsx and
  the specs into playwright/tests/*.spec.ts as plain @playwright/test
  tests, addressing stories by id and typing props via mount<typeof Story>
- drop @playwright/experimental-ct-react, vite-plugin-commonjs, the CT
  config, playwright/index.html and playwright/index.tsx
- switch to the automatic JSX runtime now that nothing needs React in
  scope, and tell Biome about it

The specs previously flaked under parallel workers (two of the core specs
failed in two runs out of three locally, masked on CI by workers: 1 and
retries: 2). Three consecutive full runs are green, and the whole suite
across all three browsers now takes ~20s.

Stories import from src/ rather than the build output, so the tests run
against sources; the published artifact is covered separately by the
package checks.
