+++ package.json
@@ -25,18 +25,17 @@
   "module": "lib/index.mjs",
   "sideEffects": false,
   "scripts": {
-    "test": "yarn test-ct",
+    "dev": "vite",
+    "test": "playwright test",
     "lint": "yarn biome check --write .",
     "build": "tsup",
-    "release": "release-it",
-    "test-ct": "playwright test -c playwright-ct.config.mts"
+    "release": "release-it"
   },
   "peerDependencies": {
     "react": "^16.8.0 || ^17.0.0 || ^18.0.0 || ^19.0.0"
   },
   "devDependencies": {
     "@biomejs/biome": "1.9.4",
-    "@playwright/experimental-ct-react": "1.62.0",
     "@playwright/test": "1.62.0",
     "@release-it/conventional-changelog": "12.0.0",
     "@types/node": "26.1.1",
@@ -49,7 +48,7 @@
     "terser": "5.44.1",
     "tsup": "8.5.1",
     "typescript": "5.9.3",
-    "vite-plugin-commonjs": "0.10.4"
+    "vite": "8.1.5"
   },
   "files": [
     "lib/"
