+++ playwright/gallery/global.d.ts
@@ -1,0 +1,11 @@
+declare global {
+  interface Window {
+    mount: (params: {
+      story: string;
+      props?: Record<string, unknown>;
+    }) => Promise<void>;
+    unmount: () => Promise<void>;
+  }
+}
+
+export {};
