+++ tsconfig.json
@@ -3,6 +3,7 @@
     "baseUrl": "./",
     "module": "esnext",
     "target": "ESNext",
+    "jsx": "react-jsx",
     "moduleResolution": "node",
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
